Tmp dir option

+++ ariba/clusters.py
@@ -1,5 +1,9 @@
+import signal
+import time
+import atexit
 import os
 import copy
+import tempfile
 import pickle
 import itertools
 import sys
@@ -13,17 +17,30 @@ from ariba import cluster, common, mapping, histogram, read_store, report, repor
 class Error (Exception): pass
 
 
-def _run_cluster(obj, verbose, clean):
+def _run_cluster(obj, verbose, clean, fails_dir):
+    failed_clusters = os.listdir(fails_dir)
+
+    if len(failed_clusters) > 0:
+        print('Other clusters failed. Stopping cluster', obj.name, file=sys.stderr)
+        return obj
+
     if verbose:
         print('Start running cluster', obj.name, 'in directory', obj.root_dir, flush=True)
-    obj.run()
+    try:
+        obj.run()
+    except:
+        print('Failed cluster:', obj.name, file=sys.stderr)
+        with open(os.path.join(fails_dir, obj.name), 'w'):
+            pass
+
     if verbose:
         print('Finished running cluster', obj.name, 'in directory', obj.root_dir, flush=True)
 
     if clean:
         if verbose:
             print('Deleting cluster dir', obj.root_dir, flush=True)
-        shutil.rmtree(obj.root_dir)
+        if os.path.exists(obj.root_dir):
+            shutil.rmtree(obj.root_dir)
 
     return obj
 
@@ -51,6 +68,7 @@ class Clusters:
       unique_threshold=0.03,
       bowtie2_preset='very-sensitive-local',
       clean=True,
+      tmp_dir=None,
     ):
         self.refdata_dir = os.path.abspath(refdata_dir)
         self.refdata, self.cluster_ids = self._load_reference_data_from_dir(refdata_dir)
@@ -64,9 +82,7 @@ class Clusters:
         else:
             self.version_report_lines = version_report_lines
 
-        self.clusters_outdir = os.path.join(self.outdir, 'Clusters')
         self.clean = clean
-
         self.logs_dir = os.path.join(self.outdir, 'Logs')
 
         self.assembler = assembler
@@ -107,12 +123,57 @@ class Clusters:
         self.clusters = {}        # gene name -> Cluster object
         self.cluster_read_counts = {} # gene name -> number of reads
         self.cluster_base_counts = {} # gene name -> number of bases
+        self.pool = None
+        self.fails_dir = os.path.join(self.outdir ,'.fails')
 
-        for d in [self.outdir, self.clusters_outdir, self.logs_dir]:
+        for d in [self.outdir, self.logs_dir, self.fails_dir]:
             try:
                 os.mkdir(d)
             except:
                 raise Error('Error mkdir ' + d)
+
+        if tmp_dir is None:
+            if 'ARIBA_TMPDIR' in os.environ:
+                tmp_dir = os.path.abspath(os.environ['ARIBA_TMPDIR'])
+            else:
+                tmp_dir = self.outdir
+
+        if not os.path.exists(tmp_dir):
+            raise Error('Temporary directory ' + tmp_dir + ' not found. Cannot continue')
+
+        self.tmp_dir = tempfile.mkdtemp(prefix='ariba.tmp.', dir=os.path.abspath(tmp_dir))
+
+        if self.verbose:
+            print('Temporary directory:', self.tmp_dir)
+
+        wanted_signals = [signal.SIGABRT, signal.SIGINT, signal.SIGSEGV, signal.SIGTERM]
+        for s in wanted_signals:
+            signal.signal(s, self._receive_signal)
+
+
+    def _stop_pool(self):
+        if self.pool is None:
+            return
+        self.pool.close()
+        self.pool.terminate()
+        while len(multiprocessing.active_children()) > 0:
+            time.sleep(1)
+
+
+    def _emergency_stop(self):
+        self._stop_pool()
+        if self.clean:
+            if os.path.exists(self.tmp_dir):
+                try:
+                    shutil.rmtree(self.tmp_dir)
+                except:
+                    pass
+
+
+    def _receive_signal(self, signum, stack):
+        print('Stopping! Signal received:', signum, file=sys.stderr, flush=True)
+        self._emergency_stop()
+        sys.exit(1)
 
 
     @classmethod
@@ -208,7 +269,7 @@ class Clusters:
 
             for ref in ref_seqs:
                 if ref not in self.cluster_to_dir:
-                    new_dir = os.path.join(self.clusters_outdir, ref)
+                    new_dir = os.path.join(self.tmp_dir, ref)
                     self.cluster_to_dir[ref] = new_dir
                     if self.verbose:
                         print('New cluster with reads that hit:', ref, flush=True)
@@ -276,7 +337,6 @@ class Clusters:
                 if self.verbose:
                     print('Constructing cluster', seq_name + '.', counter, 'of', str(len(self.cluster_to_dir)))
                 new_dir = self.cluster_to_dir[seq_name]
-                #self.refdata.write_seqs_to_fasta(os.path.join(new_dir, 'references.fa'), self.cluster_ids[seq_type][seq_name])
                 self.log_files.append(os.path.join(self.logs_dir, seq_name + '.log'))
 
                 cluster_list.append(cluster.Cluster(
@@ -285,6 +345,7 @@ class Clusters:
                     self.refdata,
                     self.cluster_read_counts[seq_name],
                     self.cluster_base_counts[seq_name],
+                    fail_file=os.path.join(self.fails_dir, seq_name),
                     read_store=self.read_store,
                     reference_names=self.cluster_ids[seq_type][seq_name],
                     logfile=self.log_files[-1],
@@ -314,11 +375,11 @@ class Clusters:
 
 
         if self.threads > 1:
-            pool = multiprocessing.Pool(self.threads)
-            cluster_list = pool.starmap(_run_cluster, zip(cluster_list, itertools.repeat(self.verbose), itertools.repeat(self.clean)))
+            self.pool = multiprocessing.Pool(self.threads)
+            cluster_list = self.pool.starmap(_run_cluster, zip(cluster_list, itertools.repeat(self.verbose), itertools.repeat(self.clean), itertools.repeat(self.fails_dir)))
         else:
             for c in cluster_list:
-                _run_cluster(c, self.verbose, self.clean)
+                _run_cluster(c, self.verbose, self.clean, self.fails_dir)
 
         self.clusters = {c.name: c for c in cluster_list}
 
@@ -367,13 +428,21 @@ class Clusters:
 
     def _clean(self):
         if self.clean:
+            shutil.rmtree(self.fails_dir)
+
             if self.verbose:
-                print('Deleting clusters directory', self.clusters_outdir)
-                shutil.rmtree(self.clusters_outdir)
+                print('Deleting tmp directory', self.tmp_dir)
+
+            if os.path.exists(self.tmp_dir):
+                shutil.rmtree(self.tmp_dir)
+
+            if self.verbose:
                 print('Deleting Logs directory', self.logs_dir)
-                shutil.rmtree(self.logs_dir)
+            shutil.rmtree(self.logs_dir)
+
+            if self.verbose:
                 print('Deleting reads store files', self.read_store.outfile + '[.tbi]')
-                self.read_store.clean()
+            self.read_store.clean()
         else:
             if self.verbose:
                 print('Not deleting anything because --noclean used')
@@ -389,6 +458,14 @@ class Clusters:
 
 
     def run(self):
+        try:
+            self._run()
+        except Error as err:
+            self._emergency_stop()
+            raise Error('Something went wrong during ariba run. Cannot continue. Error was:\n' + str(err))
+
+
+    def _run(self):
         cwd = os.getcwd()
         os.chdir(self.outdir)
         self.write_versions_file(cwd)
@@ -425,21 +502,25 @@ class Clusters:
                 print('No reads mapped. Skipping all assemblies', flush=True)
             print('WARNING: no reads mapped to reference genes. Therefore no local assemblies will be run', file=sys.stderr)
 
-        if self.verbose:
-            print('{:_^79}'.format(' Writing reports '), flush=True)
-            print('Making', self.report_file_all_tsv)
-        self._write_reports(self.clusters, self.report_file_all_tsv)
+        failed_clusters = os.listdir(self.fails_dir)
+        if len(failed_clusters):
+            print('Failed clusters:', ', '.join(failed_clusters), file=sys.stderr)
+        else:
+            if self.verbose:
+                print('{:_^79}'.format(' Writing reports '), flush=True)
+                print('Making', self.report_file_all_tsv)
+            self._write_reports(self.clusters, self.report_file_all_tsv)
 
-        if self.verbose:
-            print('Making', self.report_file_filtered_prefix + '.tsv')
-        rf = report_filter.ReportFilter(infile=self.report_file_all_tsv)
-        rf.run(self.report_file_filtered_prefix)
+            if self.verbose:
+                print('Making', self.report_file_filtered_prefix + '.tsv')
+            rf = report_filter.ReportFilter(infile=self.report_file_all_tsv)
+            rf.run(self.report_file_filtered_prefix)
 
-        if self.verbose:
-            print()
-            print('{:_^79}'.format(' Writing fasta of assembled sequences '), flush=True)
-            print(self.catted_assembled_seqs_fasta)
-        self._write_catted_assembled_seqs_fasta(self.catted_assembled_seqs_fasta)
+            if self.verbose:
+                print()
+                print('{:_^79}'.format(' Writing fasta of assembled sequences '), flush=True)
+                print(self.catted_assembled_seqs_fasta)
+            self._write_catted_assembled_seqs_fasta(self.catted_assembled_seqs_fasta)
 
         clusters_log_file = os.path.join(self.outdir, 'log.clusters.gz')
         if self.verbose:
@@ -455,5 +536,8 @@ class Clusters:
 
         if self.verbose:
             print('\nAll done!\n')
+
+        if len(failed_clusters):
+            raise Error('There were failed clusters: ' + ', '.join(failed_clusters))
 
         os.chdir(cwd)
